requirements.txt
alembic==1.8.1
click==8.1.3
Flask==2.2.2
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
python-dotenv==0.21.0
SQLAlchemy==1.4.44
Werkzeug==2.2.2
